requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==5.5.2
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
cryptography==36.0.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
Faker==18.4.0
fastjsonschema==2.15.3
fonttools==4.33.3
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
humanfriendly==10.0
idna==3.3
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.3.3
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mimesis==7.1.0
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.0
oauthlib==3.2.2
ocrmypdf==12.7.2
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20211012
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.26.1
protobuf==6.30.2
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyreadline3==3.3
pyrsistent==0.18.1
pyspnego==0.9.2
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pywinpty==2.0.5
pywinrm==0.4.3
pyzmq==23.1.0
reportlab==3.6.3
requests==2.27.1
requests-ntlm==1.2.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.62.3
traitlets==5.2.2.post1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==4.0.13
xmltodict==0.13.0
